Chapter 02 Maven project changes

+++ chapter02/Maven/Technique14/course-tracker-technique-14-end/src/main/java/com/manning/sbip/ch02/service/CourseService.java
@@ -8,12 +8,12 @@ public interface CourseService {
 
     Iterable<Course> createCourse(Course course);
 
-    Optional<Course> findCourseById(long courseId);
+    Optional<Course> findCourseById(int courseId);
 
     Iterable<Course> findAllCourses();
 
     Iterable<Course> updateCourse(Course course);
 
-    Iterable<Course> deleteCourseById(long courseId);
+    Iterable<Course> deleteCourseById(int courseId);
 
 }
